Revert "YARN-4844. Add getMemorySize/getVirtualCoresSize to o.a.h.y.api.records.Resource. (wangda)"
This reverts commit 87a641bfaba344b48cae2d2b0b46b9f486221fcd.

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/capacity/allocator/RegularContainerAllocator.java
@@ -441,7 +441,7 @@ public class RegularContainerAllocator extends AbstractContainerAllocator {
         priority, capability);
 
     // Can we allocate a container on this node?
-    long availableContainers =
+    int availableContainers =
         rc.computeAvailableContainers(available, capability);
 
     // How much need to unreserve equals to:
